feat(graph-embedding-util,senna,faba,pinto): rework projection annotation + unify annotation I/O
Refactor: split the ~1300-line type_annotation.rs into focused submodules
(markers/score/coarsen/output, alongside layout) and replace the five
`too_many_arguments` allows with context structs (InputEmbeddings,
ProjInputs, ScoreCtx, LayoutInputs).

Projection annotation (annotate-by-projection / faba gem-annotate / pinto cage):
- coarse layer reverted to Leiden community detection (k-means/type-merge
  experiments dropped)
- name communities by centered enrichment alone — drop the √|markers| weight
  that let large-marker-set types (e.g. Prog_Mk) hijack lineage names; the
  per-cell z-score is already size-matched, so the weight double-counted size
- kNN-smooth the fine per-cell scores before argmax (local consensus, default
  on); opt-in definitiveness gate (--fine-fdr / --min-margin, both off by
  default) adds a fine_margin column + 'unassigned' labels
- dedup the double Leiden: reuse the coarse communities for the co-embed
  eff-cells target (new postprocess::target_eff_from_labels)

I/O unification (both annotation methods):
- one shared write_label_tsvs (argmax.tsv + membership.tsv) and one
  finalize_annotation manifest-wiring helper, so `senna plot` auto-colours by
  either method's output identically; projection now sets manifest.annotate.argmax
- --clean (default on; --no-clean to disable) erases a prior run's output set
  first — file list owned by geu (ANNOT_OUTPUT_SUFFIXES) / an explicit
  enrichment list, never a prefix glob (the embedding/manifest are never touched)
