+++ faba/src/gem/annotate.rs
@@ -32,7 +32,9 @@
 use anyhow::{Context, Result};
 use clap::Args;
 
-use graph_embedding_util::type_annotation::{annotate_embeddings, AnnotateProjConfig};
+use graph_embedding_util::type_annotation::{
+    annotate_embeddings, AnnotateProjConfig, InputEmbeddings,
+};
 use matrix_util::common_io::mkdir_parent;
 use matrix_util::dmatrix_io::DMatrix;
 use matrix_util::traits::IoOps;
@@ -91,7 +93,7 @@ pub struct GemAnnotateArgs {
     #[arg(
         long,
         default_value_t = 30,
-        help = "k for the cell kNN graph used by the coarsening clusterer"
+        help = "k for the shared cell kNN graph (fine-score smoothing + Leiden coarsening + UMAP layout)"
     )]
     pub knn: usize,
 
@@ -179,12 +181,16 @@ pub fn run_gem_annotate(args: &GemAnnotateArgs) -> Result<()> {
         phate_max_direct: args.phate_max_direct,
         feat_knn: args.layout_knn_feat,
         umap_epochs: args.umap_epochs,
+        // Fine-score kNN smoothing + definitiveness gate use their defaults.
+        ..AnnotateProjConfig::default()
     };
     annotate_embeddings(
-        &feat.mat,
-        &feat.rows,
-        &cell.mat,
-        &cell.rows,
+        &InputEmbeddings {
+            feature_emb: &feat.mat,
+            gene_names: &feat.rows,
+            cell_emb: &cell.mat,
+            cell_names: &cell.rows,
+        },
         &args.markers,
         &format!("{out}.gem_annot"),
         !args.no_idf,
